requirements.txt
mcp>=1.0.0
httpx>=0.27.0
apscheduler>=3.10.0
pytz>=2024.1
python-dotenv>=1.0.0
